Music Queue Update
Added a queue to the music commands to allow adding endless songs to a list of music for the bot to play from. This now allows for the bot to play multiple specified songs in a row and loop songs/entire queues. Already have a list of music that you want to play? Just put all of the songs into queue!