Fixed and Improved Random
Fixed two major problems with the current random algorithm:
a) It still wasn't random enough
b) The maximum number could never be achieved

Both problems were solved and the algorithm was made to be more random by substituting Math#floor with Math#round.
This substitution is experimental and may have unintended results. Submit an issue if this occurs.